QoL and consistency changes
The run script now uses the bootstrapper as opposed to running the app
separately. This means first time setup will run in a dev environment.

The first time setup runner now has a timeout of 5 seconds to wait for
the production server to launch, and displays a loading animation while
pinging the server for a response. The page will also display an error
if the production server fails to launch, instead of redirecting to a
nonexistent web page.

The production server will now display a more useful error message if
the API URL specified does not point to a valid Augur instance. Also,
it will try an extra variation of the API URL just in case before giving
up.

Documentation for server installation now also refers to installing on a
server using Docker, and the documentation for setting up a proxy has
been separated from the systemd setup.

Logging is no longer initialized in the init startup script, but
instead in the utility script.

ServerThread is now wait/terminate compatible with subprocesses.

+++ bootstrap.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 from server import Environment
+from server import ServerThread
 import os, yaml, subprocess
 
 gunicorn_conf = """import multiprocessing
@@ -20,7 +21,6 @@ def first_time(port):
     """
     from flask import Flask, request, render_template
     from init import settings, init_settings
-    from server import ServerThread
     import threading
 
     init_settings()
@@ -147,11 +147,16 @@ if __name__ == "__main__":
         with open(config_location, "w") as config_file:
             yaml.dump(settings, config_file)
 
-    if not gunicorn_location.is_file():
+    if not env["DEVELOPMENT"] and not gunicorn_location.is_file():
         with open(gunicorn_location, "w") as gunicorn_py:
             gunicorn_py.write(gunicorn_conf)
 
-    server = subprocess.Popen(["gunicorn", "-c", str(gunicorn_location), "-b", f"{host_address}:{server_port}", "augur_view:app"])
+    if env["DEVELOPMENT"]:
+        from augur_view import app
+        server = ServerThread(app, port = server_port, address = host_address, reraise = True)
+        server.start()
+    else:
+        server = subprocess.Popen(["gunicorn", "-c", str(gunicorn_location), "-b", f"{host_address}:{server_port}", "augur_view:app"])
 
     try:
         server.wait()
